Configure for now deployments
